backend pronto al 70% van aggiunte nuove funzionalità in caso altrimenti è tranquillamente pronto per un deploy didattico

+++ workspace/Ecommerce/src/main/java/com/esempio/Ecommerce/api/dto/request/ProductRequest.java
@@ -4,7 +4,6 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 import lombok.Builder;
-import java.math.BigDecimal;
 
 @Builder
 public record ProductRequest(
@@ -15,7 +14,7 @@ public record ProductRequest(
 
         @NotNull(message = "Price is required")
         @Positive(message = "Price must be positive")
-        BigDecimal price,
+        Double price,
 
         String imageUrl,
