+++ workspace/Ecommerce/src/main/java/com/esempio/Ecommerce/api/dto/response/OrderItemResponse.java
@@ -1,15 +1,12 @@
-
 package com.esempio.Ecommerce.api.dto.response;
 
 import lombok.Builder;
-
-import java.math.BigDecimal;
 
 @Builder
 public record OrderItemResponse(
         Long productId,
         String productName,
-        BigDecimal unitPrice,
+        Double unitPrice,
         Integer quantity,
-        BigDecimal subtotal
+        Double subtotal
 ) {}
