+++ workspace/Ecommerce/src/main/java/com/esempio/Ecommerce/api/dto/response/OrderResponse.java
@@ -4,7 +4,6 @@ package com.esempio.Ecommerce.api.dto.response;
 import com.esempio.Ecommerce.domain.enums.OrderStatus;
 import lombok.Builder;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -13,7 +12,7 @@ public record OrderResponse(
         Long id,
         String userEmail,
         OrderStatus status,
-        BigDecimal total,
+        Double total,
         List<OrderItemResponse> items,
         LocalDateTime createdAt
 ) {}
