+++ workspace/Ecommerce/src/main/java/com/esempio/Ecommerce/api/dto/response/ProductResponse.java
@@ -1,7 +1,6 @@
 package com.esempio.Ecommerce.api.dto.response;
 
 import lombok.Builder;
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Builder
@@ -9,7 +8,7 @@ public record ProductResponse(
         Long id,
         String name,
         String description,
-        BigDecimal price,
+        Double price,
         String imageUrl,
         String category,
         boolean inStock,
